Cleaned main, added helper headers. Added more sensor controls. Changed program flow. restored example.hpp
